requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
python-decouple>=3.8
whitenoise>=6.5.0
dj-database-url>=2.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
bcrypt>=4.1.2
PyJWT>=2.8.0